+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import JenkinsConfiguration from './config/settings';
 import { JenkinsCodeLensProvider } from './provider/jenkins-codelens';
+import { ScriptProvider } from './provider/script-provider';
 import { XmlCodeLensProvider } from './provider/xml-codelens';
 import { BuildsProvider } from './sidebar/builds-provider';
 import { ConnectionProvider } from './sidebar/connection-provider';
@@ -10,7 +11,6 @@ import { ProjectProvider } from './sidebar/project-provider';
 import { ReservationProvider } from './sidebar/reservation-provider';
 import { SnippetProvider } from './sidebar/snippet-provider';
 import { ViewsProvider } from './sidebar/views-provider';
-import { ProjectModels } from './types/model';
 import { getConfigPath, readFileUriAsProject } from './utils/file';
 import { isRemoteUri } from './utils/remote';
 import { vscExtension } from './vsc-ns';
@@ -33,7 +33,6 @@ export async function activate(context: vscode.ExtensionContext) {
 	vscode.window.registerTreeDataProvider("utocode.views.connection", connectionProvider);
 
 	const snippetProvider = new SnippetProvider(context);
-
 	vscode.window.registerTreeDataProvider("utocode.views.reservation", reservationProvider);
 	vscode.window.registerTreeDataProvider("utocode.views.snippets", snippetProvider);
 
@@ -55,6 +54,7 @@ export async function activate(context: vscode.ExtensionContext) {
 		vscode.languages.registerCodeLensProvider(xmlLanguages, xmlLensProvider)
 	);
 
+	// const scriptProvider = new ScriptProvider(context);
 	showProjectView(projectProvider);
 	function registerCommand(cmd: string, callback: () => void) {
 		const command = vscode.commands.registerCommand(cmd, callback);
@@ -89,30 +89,8 @@ async function hasJenkinsProject(): Promise<boolean> {
 	return hasAny;
 }
 
-async function readSettings1(jenkinsSettingsPath: vscode.Uri): Promise<ProjectModels | string | undefined> {
-	if (jenkinsSettingsPath.fsPath.endsWith(".jenkinsrc.json")) {
-		if (!vscode.workspace.isTrusted) {
-			vscode.window.showInformationMessage(vscode.l10n.t("The current workspace must be Trusted in order to load settings from .jenkinsrc.json files."));
-			return undefined;
-		}
-
-		if (isRemoteUri(jenkinsSettingsPath)) {
-			vscode.window.showInformationMessage(vscode.l10n.t("This workspace contains a `.jenkinsrc.json` file, which requires the Jenkins Status extension to be installed on the remote."));
-			return undefined;
-		}
-
-		// const r = typeof __webpack_require__ === 'function' ? __non_webpack_require__ : require;
-		// delete r.cache[r.resolve(jenkinsSettingsPath.fsPath)];
-		return await jenkinsSettingsPath.fsPath;
-	} else {
-		const content = await readFileUriAsProject(jenkinsSettingsPath);
-		return content;
-	}
-}
-
 async function showProjectView(projectProvider: ProjectProvider) {
 	if (await hasJenkinsProject()) {
 		projectProvider.refresh();
 	}
 }
-